Extract useState/useEffect into custom hooks for page client components
cursor-review flagged direct useState/useEffect usage in components.
Extracts useAddConfigPageForm and useEditConfigPageForm hooks (mirroring
the old modal state hooks pattern) so the client components are pure
render wrappers.
