+++ apps/hermes/dashboard/app/dashboard/agent-configs/[id]/edit/edit-config-page-client.tsx
@@ -1,12 +1,9 @@
 "use client";
 
-import { useEffect, useMemo, useRef, useState } from "react";
-import { useRouter } from "next/navigation";
-
 import type { VariableExpansionStringFieldLoaders } from "@workspace/variable-expansion-picker";
 
-import { useFormAction } from "@/app/dashboard/agent-configs/actions/update/.generated/use-form-action";
 import { AgentConfigForm } from "../../agent-config-form";
+import { useEditConfigPageForm } from "./use-edit-config-page-form";
 
 type AgentForDropdown = {
   id: string;
@@ -30,37 +27,15 @@ type EditConfigPageClientProps = {
 };
 
 /**
- * Client wrapper for the edit config page. Manages form state and redirects on success.
+ * Client wrapper for the edit config page. Delegates state and redirect logic to useEditConfigPageForm.
  */
 export const EditConfigPageClient = ({
   config,
   agents,
   pickerLoaders,
 }: EditConfigPageClientProps) => {
-  const router = useRouter();
-  const [formState, setFormState] = useState({
-    name: config.name,
-    description: config.description ?? "",
-    agentKey: `${config.agentId}@${config.agentVersion}`,
-    config:
-      typeof config.config === "object" && config.config !== null
-        ? { ...(config.config as Record<string, unknown>) }
-        : {},
-  });
-  const { FormWithAction, state, pending } = useFormAction();
-  const wasPendingRef = useRef(false);
-
-  const errorMessage = useMemo(() => {
-    if (state && state.status === false) return state.message as string;
-    return null;
-  }, [state]);
-
-  useEffect(() => {
-    if (wasPendingRef.current && !pending && state?.status === true) {
-      router.push("/dashboard/agent-configs");
-    }
-    wasPendingRef.current = pending;
-  }, [pending, state, router]);
+  const { formState, setFormState, FormWithAction, pending, errorMessage } =
+    useEditConfigPageForm(config);
 
   return (
     <AgentConfigForm
